scylla
Module client

This module holds entities that represent the whole configurable driver’s client of the cluster. The following abstractions are involved:

Session - the main entity of the driver. It:

contains and manages all driver configuration,

enables executing CQL requests, taking all configuration into consideration.

SessionBuilder - just a convenient builder for a Session.

CachingSession - a wrapper over a Session that keeps and manages a cache of prepared statements, so that a user can be free of such considerations.

SelfIdentity - configuresd driver and application self-identifying information, to be sent in STARTUP message.

ExecutionProfile - a profile that groups various configuration options relevant when executing a request against the DB.

QueryPager and TypedRowStream - entities that provide automated transparent paging of a query.

Modules

caching_session

execution_profile

ExecutionProfile is a grouping of configurable options regarding CQL statement execution.

pager

Entities that provide automated transparent paging of a query. They enable consuming result of a paged query as a stream over rows, which abstracts over page boundaries.

session

Session is the main object used in the driver.
It manages all connections to the cluster and allows to execute CQL requests.

session_builder

SessionBuilder provides an easy way to create new Sessions

Structs

SelfIdentity

Driver and application self-identifying information, to be sent in STARTUP message.

Enums

Compression

The wire protocol compression algorithm.

PoolSize

The target size of a per-node connection pool.

WriteCoalescingDelay

Represents a write coalescing delay configuration option.