Module scylla::deserialize::value

Deserializing a single CQL value from a column of the query result row.

Structs

BuiltinDeserializationError

Deserialization of one of the built-in types failed.

BuiltinTypeCheckError

Type checking of one of the built-in types failed.

ListlikeIterator

An iterator over either a CQL set or list.

MapIterator

An iterator over a CQL map.

UdtIterator

An iterator over fields of a User Defined Type.

Enums

BuiltinDeserializationErrorKind

Describes why deserialization of some of the built-in types failed.

BuiltinTypeCheckErrorKind

Describes why type checking some of the built-in types failed.

MapDeserializationErrorKind

Describes why deserialization of a map type failed.

MapTypeCheckErrorKind

Describes why type checking of a map type failed.

MaybeEmpty

A value that may be empty or not.

SetOrListDeserializationErrorKind

Describes why deserialization of a set or list type failed.

SetOrListTypeCheckErrorKind

Describes why type checking of a set or list type failed.

TupleDeserializationErrorKind

Describes why deserialization of a tuple failed.

TupleTypeCheckErrorKind

Describes why type checking of a tuple failed.

UdtTypeCheckErrorKind

Describes why type checking of a user defined type failed.

Traits

Emptiable

Values that may be empty or not.