scylla::deserialize::value

Trait Emptiable

pub trait Emptiable { }

Values that may be empty or not.

In CQL, some types can have a special value of “empty”, represented as a serialized value of length 0. An example of this are integral types: the “int” type can actually hold 2^32 + 1 possible values because of this quirk. Note that this is distinct from being NULL.

Rust types that cannot represent an empty value (e.g. i32) should implement this trait in order to be deserialized as MaybeEmpty.