scylla::errors

Enum RequestAttemptError

#[non_exhaustive]
pub enum RequestAttemptError {
    SerializationError(SerializationError),
    CqlRequestSerialization(CqlRequestSerializationError),
    UnableToAllocStreamId,
    BrokenConnectionError(BrokenConnectionError),
    BodyExtensionsParseError(FrameBodyExtensionsParseError),
    CqlResultParseError(CqlResultParseError),
    CqlErrorParseError(CqlErrorParseError),
    DbError(DbError, String),
    UnexpectedResponse(CqlResponseKind),
    RepreparedIdChanged {
        statement: String,
        expected_id: Vec<u8>,
        reprepared_id: Vec<u8>,
    },
    RepreparedIdMissingInBatch,
    NonfinishedPagingState,
}

An error that occurred during a single attempt of:

QUERY

PREPARE

EXECUTE

BATCH

requests. The retry decision is made based on this error.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

SerializationError(SerializationError)

Failed to serialize query parameters. This error occurs, when user executes a CQL QUERY request with non-empty parameterโs value list and the serialization of provided values fails during statement preparation.

CqlRequestSerialization(CqlRequestSerializationError)

Failed to serialize CQL request.

UnableToAllocStreamId

Driver was unable to allocate a stream id to execute a query on.

BrokenConnectionError(BrokenConnectionError)

A connection has been broken during query execution.

BodyExtensionsParseError(FrameBodyExtensionsParseError)

Failed to deserialize frame body extensions.

CqlResultParseError(CqlResultParseError)

Received a RESULT server response, but failed to deserialize it.

CqlErrorParseError(CqlErrorParseError)

Received an ERROR server response, but failed to deserialize it.

DbError(DbError, String)

Database sent a response containing some error with a message

UnexpectedResponse(CqlResponseKind)

Received an unexpected response from the server.

RepreparedIdChanged

Prepared statement id changed after repreparation.

Fields

statement: String

expected_id: Vec<u8>

reprepared_id: Vec<u8>

RepreparedIdMissingInBatch

Driver tried to reprepare a statement in the batch, but the reprepared statementโs id is not included in the batch.

NonfinishedPagingState

A result with nonfinished paging state received for unpaged query.