scylla::frame::response

Module result

Structs

ColumnSpec

Specification of a column of a table.

PartitionKeyIndex

Represents the relationship between partition key columns and bind markers. This allows implementations with token-aware routing to correctly construct the partition key without needing to inspect table metadata.

TableSpec

Specification of a table in a keyspace.

UserDefinedType

Definition of a user-defined type (UDT). UDT is composed of fields, each with a name and an optional value of its own type.

Enums

CollectionType

Collection variants of ColumnType. A collection is a composite type that has dynamic size, so it is possible to add and remove values to/from it.

ColumnType

A type of:

NativeType

A ColumnType variants that are “simple” (non-recursive).