Enum scylla::frame::response::result::CqlValue

pub enum CqlValue {

    Ascii(String),
    Boolean(bool),
    Blob(Vec<u8>),
    Counter(Counter),
    Decimal(CqlDecimal),
    Date(CqlDate),
    Double(f64),
    Duration(CqlDuration),
    Empty,
    Float(f32),
    Int(i32),
    BigInt(i64),
    Text(String),
    Timestamp(CqlTimestamp),
    Inet(IpAddr),
    List(Vec<CqlValue>),
    Map(Vec<(CqlValue, CqlValue)>),
    Set(Vec<CqlValue>),
    UserDefinedType {
        keyspace: String,
        type_name: String,
        fields: Vec<(String, Option<CqlValue>)>,
    },
    SmallInt(i16),
    TinyInt(i8),
    Time(CqlTime),
    Timeuuid(CqlTimeuuid),
    Tuple(Vec<Option<CqlValue>>),
    Uuid(Uuid),
    Varint(CqlVarint),

}

Variants

Ascii(String)

Boolean(bool)

Blob(Vec<u8>)

Counter(Counter)

Decimal(CqlDecimal)

Date(CqlDate)

Days since -5877641-06-23 i.e. 2^31 days before unix epoch Can be converted to chrono::NaiveDate (-262145-1-1 to 262143-12-31) using as_date

Double(f64)

Duration(CqlDuration)

Empty

Float(f32)

Int(i32)

BigInt(i64)

Text(String)

Timestamp(CqlTimestamp)

Milliseconds since unix epoch

Inet(IpAddr)

List(Vec<CqlValue>)

Map(Vec<(CqlValue, CqlValue)>)

Set(Vec<CqlValue>)

UserDefinedType

Fields

keyspace: String

type_name: String

fields: Vec<(String, Option<CqlValue>)>

Order of fields vector must match the order of fields as defined in the UDT. The driver does not check it by itself, so incorrect data will be written if the order is wrong.

SmallInt(i16)

TinyInt(i8)

Time(CqlTime)

Nanoseconds since midnight

Timeuuid(CqlTimeuuid)

Tuple(Vec<Option<CqlValue>>)

Uuid(Uuid)

Varint(CqlVarint)

Implementations

impl CqlValue

pub fn as_ascii(&self) -> Option<&String>

pub fn as_cql_date(&self) -> Option<CqlDate>

pub fn as_cql_timestamp(&self) -> Option<CqlTimestamp>

pub fn as_cql_time(&self) -> Option<CqlTime>

pub fn as_cql_duration(&self) -> Option<CqlDuration>

pub fn as_counter(&self) -> Option<Counter>

pub fn as_boolean(&self) -> Option<bool>

pub fn as_double(&self) -> Option<f64>

pub fn as_uuid(&self) -> Option<Uuid>

pub fn as_float(&self) -> Option<f32>

pub fn as_int(&self) -> Option<i32>

pub fn as_bigint(&self) -> Option<i64>

pub fn as_tinyint(&self) -> Option<i8>

pub fn as_smallint(&self) -> Option<i16>

pub fn as_blob(&self) -> Option<&Vec<u8>>

pub fn as_text(&self) -> Option<&String>

pub fn as_timeuuid(&self) -> Option<CqlTimeuuid>

pub fn into_string(self) -> Option<String>

pub fn into_blob(self) -> Option<Vec<u8>>

pub fn as_inet(&self) -> Option<IpAddr>

pub fn as_list(&self) -> Option<&Vec<CqlValue>>

pub fn as_set(&self) -> Option<&Vec<CqlValue>>

pub fn as_map(&self) -> Option<&Vec<(CqlValue, CqlValue)>>

pub fn as_udt(&self) -> Option<&Vec<(String, Option<CqlValue>)>>

pub fn into_vec(self) -> Option<Vec<CqlValue>>

pub fn into_pair_vec(self) -> Option<Vec<(CqlValue, CqlValue)>>

pub fn into_udt_pair_vec(self) -> Option<Vec<(String, Option<CqlValue>)>>

pub fn into_cql_varint(self) -> Option<CqlVarint>

pub fn into_cql_decimal(self) -> Option<CqlDecimal>