scylla::frame

Module types

Enums

Consistency

A setting that defines a successful write or read by the number of cluster replicas that acknowledge the write or respond to the read request, respectively. See ScyllaDB docs for more detailed description and guidelines.

SerialConsistency

Consistency for Lightweight Transactions (LWTs).