Module scylla::frame::value

Structs

SerializedValues
Keeps a buffer with serialized Values Allows adding new Values and iterating over serialized ones

SerializedValuesIterator

ValueTooBig

Enums

SerializeValuesError

Traits

BatchValues
Represents List of ValueList for Batch statement

Value
Every value being sent in a query must implement this trait serialize() should write the Value as bytes to the provided buffer

ValueList
Represents list of values to be sent in a query gets serialized and but into request

Type Definitions

SerializedResult