Trait scylla::frame::value::LegacyBatchValues

pub trait LegacyBatchValues {
    type LegacyBatchValuesIter<'r>: LegacyBatchValuesIterator<'r>
       where Self: 'r;

    // Required method
    fn batch_values_iter(&self) -> Self::LegacyBatchValuesIter<'_>;
}

Represents List of ValueList for Batch statement

Required Associated Types

type LegacyBatchValuesIter<'r>: LegacyBatchValuesIterator<'r> where Self: 'r

For some unknown reason, this type, when not resolved to a concrete type for a given async function, cannot live across await boundaries while maintaining the corresponding future Send, unless 'r: 'static

See https://github.com/scylladb/scylla-rust-driver/issues/599 for more details

Required Methods

fn batch_values_iter(&self) -> Self::LegacyBatchValuesIter<'_>

Object Safety

This trait is not object safe.