Trait scylla::frame::value::Value

pub trait Value {
    // Required method
    fn serialize(&self, buf: &mut Vec<u8>) -> Result<(), ValueTooBig>;
}

Every value being sent in a query must implement this trait serialize() should write the Value as bytes to the provided buffer

Required Methods

fn serialize(&self, buf: &mut Vec<u8>) -> Result<(), ValueTooBig>