scylla
Module history

Collecting history of query executions - retries, speculative, etc.

Structs

AttemptHistory

AttemptId

Id of a single attempt within a query, a single request sent on some connection.

FiberHistory

HistoryCollector

HistoryCollector can be used as HistoryListener to collect all the query history events. Each event is marked with an UTC timestamp.

HistoryCollectorData

QueryHistory

QueryId

Id of a single query, i.e. a single call to Session::{query,execute}_{unpaged,single_page}/etc.

SpeculativeId

Id of a speculative execution fiber. When speculative execution is enabled the driver will start multiple speculative threads, each of them performing sequential attempts.

StructuredHistory

Structured representation of queries history.
HistoryCollector collects raw events which later can be converted to this pretty representation.
It has a Display impl which can be used for printing pretty query history.

Enums

AttemptResult

HistoryEvent

QueryHistoryResult

Traits

HistoryListener

Any type implementing this trait can be passed to Session to collect execution history of specific queries.
In order to use it call set_history_listener on Query, PreparedStatement, etc…
The listener has to generate unique IDs for new queries, attempts and speculative fibers. These ids are then used by the caller to identify them.
It’s important to note that even after a query is finished there still might come events related to it. These events come from speculative futures that didn’t notice the query is done already.

Type Aliases

TimePoint