Module scylla::macros

Structs

Bytes
A cheaply cloneable and sliceable chunk of contiguous memory.

BytesMut
A unique reference to a contiguous slice of memory.

Traits

BufMut
A trait for values that provide sequential write access to bytes.

Derive Macros

FromRow
#[derive(FromRow)] derives FromRow for struct Works only on simple structs without generics etc

FromUserType
#[derive(FromUserType)] allows to parse struct as a User Defined Type Works only on simple structs without generics etc

IntoUserType
#[derive(IntoUserType)] allows to pass struct a User Defined Type Value in queries Works only on simple structs without generics etc