scylla::policies
Module retry

Structs

DefaultRetryPolicy

Default retry policy - retries when there is a high chance that a retry might help.
Behaviour based on DataStax Java Driver

DefaultRetrySession

DowngradingConsistencyRetryPolicy

Downgrading consistency retry policy - retries with lower consistency level if it knows
that the initial CL is unreachable. Also, it behaves as DefaultRetryPolicy when it believes that the initial CL is reachable. Behaviour based on [DataStax Java Driver]
(https://docs.datastax.com/en/drivers/java/3.11/com/datastax/driver/core/policies/DowngradingConsistencyRetryPolicy.html)

DowngradingConsistencyRetrySession

FallthroughRetryPolicy

Forwards all errors directly to the user, never retries

FallthroughRetrySession

RequestInfo

Information about a failed request

Enums

RetryDecision

Returned by implementations of RetryPolicy. Instructs the driver on what to do about the request after it failed.

Traits

RetryPolicy

Specifies a policy used to decide when to retry a request

RetrySession

Used throughout a single request to decide when to retry it After this request is finished it is destroyed or reset