scylla::policies::retry

Trait RetryPolicy

pub trait RetryPolicy:
    Debug
    + Send
    + Sync {
    // Required method
    fn new_session(&self) -> Box<dyn RetrySession>;
}

Specifies a policy used to decide when to retry a request

Required Methods

fn new_session(&self) -> Box<dyn RetrySession>

Called for each new request, starts a session of deciding about retries