scylla::policies::retry

Trait RetrySession

pub trait RetrySession: Send + Sync {
    // Required methods
    fn decide_should_retry(
        &mut self,
        request_info: RequestInfo<'_>,
    ) -> RetryDecision;
    fn reset(&mut self);
}

Used throughout a single request to decide when to retry it After this request is finished it is destroyed or reset

Required Methods

fn decide_should_retry( &mut self, request_info: RequestInfo<'_>, ) -> RetryDecision

Called after the request failed - decide what to do next

fn reset(&mut self)

Reset before using for a new request