scylla
Module serialize

Serializing bound values of a query to be sent to the DB.

Modules

batch

Contains the BatchValues and BatchValuesIterator trait and their implementations.

row

Contains the SerializeRow trait and its implementations.

value

Contains the SerializeValue trait and its implementations.

writers

Contains types and traits used for safe serialization of values for a CQL statement.

Structs

SerializationError

An error indicating that a failure happened during serialization.