scylla::serialize
Module batch

Contains the BatchValues and BatchValuesIterator trait and their implementations.

Structs

BatchValuesFromIterator

Implements BatchValues from an Iterator over references to things that implement SerializeRow

BatchValuesIteratorFromIterator

Implements BatchValuesIterator from an Iterator over references to things that implement SerializeRow

LegacyBatchValuesAdapterDeprecated

A newtype wrapper which adjusts an existing types that implement LegacyBatchValues to the current BatchValues API.

LegacyBatchValuesIteratorAdapterDeprecated

A newtype wrapper which adjusts an existing types that implement LegacyBatchValuesIterator to the current BatchValuesIterator API.

TupleValuesIter

A BatchValuesIterator over a tuple.

Traits

BatchValues

Represents a list of sets of values for a batch statement.

BatchValuesIterator

An Iterator-like object over the values from the parent BatchValues object.