scylla::serialize::batch

Trait BatchValues

pub trait BatchValues {
    type BatchValuesIter<'r>: BatchValuesIterator<'r>
       where Self: 'r;

    // Required method
    fn batch_values_iter(&self) -> Self::BatchValuesIter<'_>;
}

Represents a list of sets of values for a batch statement.

The data in the object can be consumed with an iterator-like object returned by the BatchValues::batch_values_iter method.

Required Associated Types

type BatchValuesIter<'r>: BatchValuesIterator<'r> where Self: 'r

An Iterator-like object over the values from the parent BatchValues object.

Required Methods

fn batch_values_iter(&self) -> Self::BatchValuesIter<'_>

Returns an iterator over the data contained in this object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.