scylla::serialize
Module row

Contains the SerializeRow trait and its implementations.

Structs

BuiltinSerializationError

Failed to serialize values for a statement, represented by one of the types built into the driver.

BuiltinTypeCheckError

Failed to type check values for a statement, represented by one of the types built into the driver.

RowSerializationContext

Contains information needed to serialize a row.

SerializedValues

A buffer containing already serialized values.

SerializedValuesIterator

An iterator over raw values in some SerializedValues.

ValueListAdapterDeprecated

Implements SerializeRow if the type wrapped over implements ValueList.

Enums

BuiltinSerializationErrorKind

Describes why serializing values for a statement failed.

BuiltinTypeCheckErrorKind

Describes why type checking values for a statement failed.

ValueListToSerializeRowAdapterErrorDeprecated

Describes a failure to translate the output of the ValueList legacy trait into an output of the SerializeRow trait.

Traits

SerializeRow

Represents a set of values that can be sent along a CQL statement.

Functions

serialize_legacy_rowDeprecated

Serializes an object implementing ValueList by using the RowWriter interface.