scylla

Module statement

This module holds entities representing various kinds of CQL statements, together with their execution options. The following statements are supported:

Query (unprepared statements),

PreparedStatement,

Batch.

Re-exports

pub use unprepared::Statement;

Modules

batch

prepared

unprepared

Enums

Consistency

A setting that defines a successful write or read by the number of cluster replicas that acknowledge the write or respond to the read request, respectively. See ScyllaDB docs for more detailed description and guidelines.

SerialConsistency

Consistency for Lightweight Transactions (LWTs).