scylla::statement::unprepared

Struct Statement

pub struct Statement {
    pub contents: String,
    /* private fields */
}

Unprepared CQL statement.

This represents a CQL statement that can be executed on a server.

Fields
contents: String
Implementations

impl Statement

pub fn new(query_text: impl Into<String>) -> Self

Creates a new Statement from a CQL statement string.

pub fn with_page_size(self, page_size: i32) -> Self

Returns self with page size set to the given value.

Panics if given number is nonpositive.

pub fn set_page_size(&mut self, page_size: i32)

Sets the page size for this CQL statement.

Panics if given number is nonpositive.

pub fn get_page_size(&self) -> i32

Returns the page size for this CQL statement.

pub fn set_consistency(&mut self, c: Consistency)

Sets the consistency to be used when executing this statement.

pub fn get_consistency(&self) -> Option<Consistency>

Gets the consistency to be used when executing this statement if it is filled. If this is empty, the default_consistency of the session will be used.

pub fn set_serial_consistency(&mut self, sc: Option<SerialConsistency>)

Sets the serial consistency to be used when executing this statement. (Ignored unless the statement is an LWT)

pub fn get_serial_consistency(&self) -> Option<SerialConsistency>

Gets the serial consistency to be used when executing this statement. (Ignored unless the statement is an LWT)

pub fn set_is_idempotent(&mut self, is_idempotent: bool)

Sets the idempotence of this statement A statement is idempotent if it can be applied multiple times without changing the result of the initial application If set to true we can be sure that it is idempotent If set to false it is unknown whether it is idempotent This is used in RetryPolicy to decide if retrying a statement execution is safe

pub fn get_is_idempotent(&self) -> bool

Gets the idempotence of this statement

pub fn set_tracing(&mut self, should_trace: bool)

Enable or disable CQL Tracing for this statement If enabled session.query() will return a QueryResult containing tracing_id which can be used to query tracing information about the execution of this query

pub fn get_tracing(&self) -> bool

Gets whether tracing is enabled for this statement

pub fn set_timestamp(&mut self, timestamp: Option<i64>)

Sets the default timestamp for this statement in microseconds. If not None, it will replace the server side assigned timestamp as default timestamp If a statement contains a USING TIMESTAMP clause, calling this method wonโt change anything

pub fn get_timestamp(&self) -> Option<i64>

Gets the default timestamp for this statement in microseconds.

pub fn set_request_timeout(&mut self, timeout: Option<Duration>)

Sets the client-side timeout for this statement. If not None, the driver will stop waiting for the request to finish after timeout passed. Otherwise, default session client timeout will be applied.

pub fn get_request_timeout(&self) -> Option<Duration>

Gets client timeout associated with this statement.

pub fn set_retry_policy(&mut self, retry_policy: Option<Arc<dyn RetryPolicy>>)

Set the retry policy for this statement, overriding the one from execution profile if not None.

pub fn get_retry_policy(&self) -> Option<&Arc<dyn RetryPolicy>>

Get the retry policy set for the statement.

pub fn set_history_listener( &mut self, history_listener: Arc<dyn HistoryListener>, )

Sets the listener capable of listening what happens during statement execution.

pub fn remove_history_listener(&mut self) -> Option<Arc<dyn HistoryListener>>

Removes the listener set by set_history_listener.

pub fn set_execution_profile_handle( &mut self, profile_handle: Option<ExecutionProfileHandle>, )

Associates the query with execution profile referred by the provided handle. Handle may be later remapped to another profile, and query will reflect those changes.

pub fn get_execution_profile_handle(&self) -> Option<&ExecutionProfileHandle>

Borrows the execution profile handle associated with this statement.