scylla::transport
Module errors

This module contains various errors which can be returned by Session.

Structs

BrokenConnectionError

An error indicating that a connection was broken. Possible error reasons:

ConnectionSetupRequestError

An error that occurred during connection setup request execution. It indicates that request needed to initiate a connection failed.

Enums

BadKeyspaceName

Invalid keyspace name given to Session::use_keyspace()

BadQuery

Error caused by caller creating an invalid query

BrokenConnectionErrorKind

A reason why connection was broken.

ConnectionError

An error that appeared on a connection level. It indicated that connection can no longer be used and should be dropped.

ConnectionPoolError

An error that occurred when selecting a node connection to perform a request on.

ConnectionSetupRequestErrorKind

CqlEventHandlingError

Failed to handle a CQL event received on a stream -1. Possible error kinds are:

DbError

An error sent from the database in response to a query as described in the specification\

KeyspaceStrategyError

An error that occurred during specific keyspace’s metadata fetch.

KeyspacesMetadataError

An error that occurred during keyspaces metadata fetch.

MetadataError

An error that occurred during cluster metadata fetch.

NewSessionError

Error that occurred during session creation

OperationType

Type of the operation rejected by rate limiting

PeersMetadataError

An error that occurred during peers metadata fetch.

ProtocolError

A protocol error.

QueryError

Error that occurred during query execution

RequestError

An error that occurred when performing a request.

SchemaVersionFetchError

A protocol error that occurred during schema version fetch.

TablesMetadataError

An error that occurred during tables metadata fetch.

TracingProtocolError

A protocol error that occurred during tracing info fetch.

TranslationError

Error caused by failed address translation done before establishing connection

UdtMetadataError

An error that occurred during UDTs metadata fetch.

UseKeyspaceProtocolError

A protocol error that occurred during USE KEYSPACE <> request.

ViewsMetadataError

An error that occurred during views metadata fetch.

WriteType

Type of write operation requested