scylla::transport
Module iterator

Iterators over rows returned by paged queries

Structs

LegacyRowIteratorDeprecated

Iterator over rows returned by paged queries.

LegacyTypedRowIteratorDeprecated

Iterator over rows returned by paged queries where each row is parsed as the given type
Returned by RowIterator::into_typed

QueryPager

An intermediate object that allows to construct a stream over a query that is asynchronously paged in the background.

TypedRowStream

Returned by QueryPager::rows_stream.

Enums

LegacyNextRowErrorDeprecated

Couldn’t get next typed row from the iterator

NextPageError

An error returned that occurred during next page fetch.

NextRowError

An error returned by async iterator API.