scylla::transport
Module legacy_query_result

Structs

LegacyQueryResultDeprecated

Result of a single query
Contains all rows returned by the database and some more information

RowsExpectedErrorDeprecated

LegacyQueryResult::rows() or a similar function called on a bad LegacyQueryResult.
Expected LegacyQueryResult.rows to be Some, but it was None.
LegacyQueryResult.rows is Some for queries that can return rows (e.g SELECT).
It is None for queries that can’t return rows (e.g INSERT).

RowsNotExpectedError

LegacyQueryResult::result_not_rows() called on a bad LegacyQueryResult.
Expected LegacyQueryResult.rows to be None, but it was Some.
LegacyQueryResult.rows is Some for queries that can return rows (e.g SELECT).
It is None for queries that can’t return rows (e.g INSERT).

TypedRowIterDeprecated

Iterator over rows parsed as the given type
Returned by rows.into_typed::<(...)>()

Enums

FirstRowError

FirstRowTypedError

IntoLegacyQueryResultErrorDeprecated

An error that occurred during QueryResult to LegacyQueryResult conversion.

MaybeFirstRowTypedErrorDeprecated

SingleRowErrorDeprecated

SingleRowTypedErrorDeprecated

Traits

IntoTypedRowsDeprecated

Trait used to implement Vec<result::Row>::into_typed<RowT>