scylla::transport::legacy_query_result
Trait IntoTypedRows

pub trait IntoTypedRows {
    // Required method
    fn into_typed<RowT: FromRow>(self) -> TypedRowIter<RowT> ;
}

👎Deprecated since 0.15.1: Legacy deserialization API is inefficient and is going to be removed soon

Trait used to implement Vec<result::Row>::into_typed<RowT>

Required Methods

fn into_typed<RowT: FromRow>(self) -> TypedRowIter<RowT>

👎Deprecated since 0.15.1: Legacy deserialization API is inefficient and is going to be removed soon

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.