Module scylla::transport::load_balancing

Structs

DCAwareRoundRobinPolicy
A data-center aware Round-robin load balancing policy.

RoundRobinPolicy
A Round-robin load balancing policy.

Statement
Represents info about statement that can be used by load balancing policies.

TokenAwarePolicy
A wrapper load balancing policy that adds token awareness to a child policy.

Traits

ChildLoadBalancingPolicy
This trait is used to apply policy to plan made by parent policy.

LoadBalancingPolicy
Policy that decides which nodes to contact for each query