Trait scylla::transport::load_balancing::LoadBalancingPolicy

pub trait LoadBalancingPolicy: Send + Sync {
    fn plan<'a>(
        &self, 
        statement: &Statement, 
        cluster: &'a ClusterData
    ) -> Box<dyn Iterator<Item = Arc<Node>> + 'a>;

    fn name(&self) -> String;
}

Policy that decides which nodes to contact for each query

Required methods

fn plan<'a>(
&self,
statement: &Statement,
cluster: &'a ClusterData
) -> Box<dyn Iterator<Item = Arc<Node>> + 'a>

It is used for each query to find which nodes to query first

fn name(&self) -> String

Returns name of load balancing policy

Loading content...