Module scylla::transport::locator

Structs

ReplicaLocator

ReplicaLocator provides a way to find the set of owning nodes for a given (token, replication strategy) pair. It does so by either using the precomputed token ranges, or doing the computation on the fly.

ReplicaSet

Represents a set of replicas for a given token and strategy;

ReplicaSetIterator

Iterator that returns replicas from some replica set.

ReplicasOrdered

Represents a sequence of replicas for a given token and strategy, ordered according to the ring order.

ReplicasOrderedIterator

Iterator that returns replicas from some replica sequence, ordered according to the ring order.

TokenRing

A token ring is a continuous hash ring. It defines association by hashing a key onto the ring and then walking the ring in one direction. Cassandra and Scylla use it for determining data ownership which allows for efficient load balancing. The token ring is used by the driver to find the replicas for a given token. Each ring member has a token (i64 number) which defines the member’s position on the ring. The ring is circular and can be traversed in the order of increasing tokens. TokenRing makes it easy and efficient to traverse the ring starting at a given token.