Struct scylla::transport::locator::ReplicaSet

pub struct ReplicaSet<'a> { /* private fields */ }

Represents a set of replicas for a given token and strategy;

This container can only be created by calling ReplicaLocator::replicas_for_token, and it can borrow precomputed replica lists living in the locator.

Implementations

impl<'a> ReplicaSet<'a>

pub fn choose_filtered<R>( self, rng: &mut R, predicate: impl Fn(&NodeRef<'a>) -> bool, ) -> Option<NodeRef<'a>>
where R: Rng + ?Sized,

Chooses a random replica that satisfies the given predicate.

pub fn len(&self) -> usize

Gets the size of the set.

If the set represents SimpleStrategy replicas that were filtered by datacenter, this function will have O(R) complexity, where R is the replication factor of that strategy.

If the set represents NetworkTopologyStrategy replicas that were not filtered by datacenter, this function will have O(D) complexity where D is the number of known datacenters.

In all other cases, the complexity is O(1)

pub fn is_empty(&self) -> bool

Returns true if the replica set contains no elements.

Complexity same as of ReplicaSet::len.

impl<'a> ReplicaSet<'a>

pub fn into_replicas_ordered(self) -> ReplicasOrdered<'a>