scylla::transport::partitioner
Trait PartitionerHasher

pub trait PartitionerHasher {
    // Required methods
    fn write(&mut self, pk_part: &[u8]);
    fn finish(&self) -> Token;
}

A trait for hashing a stream of serialized CQL values.

Instances of this trait are created by a Partitioner and are stateful. At any point, one can call finish() and a Token will be computed based on values that has been fed so far.

Required Methods

fn write(&mut self, pk_part: &[u8])

fn finish(&self) -> Token