scylla::transport
Module query_result

Structs

ColumnSpecView

A view over specification of a column returned by the database.

ColumnSpecs

A view over specification of columns returned by the database.

QueryResult

Result of a single request to the database. It represents any kind of Result frame.

QueryRowsResult

Enables deserialization of rows received from the database in a QueryResult.

ResultNotRowsError

An error returned by QueryResult::result_not_rows.

TableSpecView

A view over specification of a table in the database.

Enums

FirstRowError

An error returned by QueryRowsResult::first_row.

IntoRowsResultError

An error returned by QueryResult::into_rows_result

MaybeFirstRowError

An error returned by QueryRowsResult::maybe_first_row.

RowsError

An error returned by QueryRowsResult::rows.

SingleRowError

An error returned by QueryRowsResult::single_row.