scylla::transport::retry_policy
Trait RetrySession

pub trait RetrySession: Send + Sync {
    // Required methods
    fn decide_should_retry(
        &mut self,
        query_info: QueryInfo<'_>,
    ) -> RetryDecision;
    fn reset(&mut self);
}

Used throughout a single query to decide when to retry it After this query is finished it is destroyed or reset

Required Methods

fn decide_should_retry(&mut self, query_info: QueryInfo<'_>) -> RetryDecision

Called after the query failed - decide what to do next

fn reset(&mut self)

Reset before using for a new query