scylla::transport
Module session

Session is the main object used in the driver.
It manages all connections to the cluster and allows to perform queries.

Re-exports

pub use crate::transport::errors::TranslationError;

pub use crate::transport::legacy_query_result::IntoTypedRows;

Deprecated

pub use crate::transport::legacy_query_result::TypedRowIter;

Deprecated

Structs

GenericSession

Session manages connections to the cluster and allows to perform queries

SessionConfig

Configuration options for Session. Can be created manually, but usually it’s easier to use SessionBuilder

Enums

CurrentDeserializationApi

LegacyDeserializationApiDeprecated

PoolSize

The target size of a per-node connection pool.

Traits

AddressTranslator

Translates IP addresses received from ScyllaDB nodes into locally reachable addresses.

DeserializationApiKind

Type Aliases

LegacySessionDeprecated

Session