Struct scylla::transport::session::Session

pub struct Session { /* fields omitted */ }

Implementations

impl Session

Represents a CQL session, which can be used to communicate with the database

pub async fn connect(config: SessionConfig) -> Result<Self, NewSessionError>

Estabilishes a CQL session with the database

Arguments

config - Connectiong configuration - known nodes, Compression, etc.

pub async fn query(
&self,
query: impl Into<Query>,
values: impl ValueList
) -> Result<Option<Vec<Row>>, QueryError>

Sends a query to the database and receives a response. If query has paging enabled, this will return only the first page.

Arguments

query - query to be performed

values - values bound to the query

pub async fn query_iter(
&self,
query: impl Into<Query>,
values: impl ValueList
) -> Result<RowIterator, QueryError>

pub async fn prepare(
&self,
query: &str
) -> Result<PreparedStatement, QueryError>

Prepares a statement on the server side and returns a prepared statement, which can later be used to perform more efficient queries

Arguments

pub async fn execute(
&self,
prepared: &PreparedStatement,
values: impl ValueList
) -> Result<Option<Vec<Row>>, QueryError>

Executes a previously prepared statement

Arguments

prepared - a statement prepared with prepare

values - values bound to the query

pub async fn execute_iter(
&self,
prepared: impl Into<PreparedStatement>,
values: impl ValueList
) -> Result<RowIterator, QueryError>

pub async fn batch(
&self,
batch: &Batch,
values: impl BatchValues
) -> Result<(), QueryError>

Sends a batch to the database.

Arguments

batch - batch to be performed

values - values bound to the query

pub async fn refresh_topology(&self) -> Result<(), QueryError>

pub fn get_metrics(&self) -> MetricsView