Crate scylla

Async Rust driver for the Scylla database written in Rust. Although optimized for Scylla, the driver is also compatible with Apache Cassandra®.

Documentation book

The best source to learn about this driver is the documentation book.
This page contains mainly API documentation

Other documentation

Documentation book

Examples

Scylla documentation

Cassandra® documentation

Driver overview

Connecting

All driver activity revolves around the Session
Session is created by specifying a few known nodes and connecting to them:

use scylla::client::session::Session;
use scylla::client::session_builder::SessionBuilder;
use std::error::Error;

#[tokio::main]
async fn main() -> Result<(), Box<dyn Error>> {
   let session: Session = SessionBuilder::new()
        .known_node("127.0.0.1:9042")
        .known_node("1.2.3.4:9876")
        .build()
        .await?;

   Ok(())
}

Session is usually created using the SessionBuilder.
All configuration options for a Session can be specified while building.

Making queries

After successfully connecting to the cluster we can make queries.
The driver supports multiple query types:

Simple

Simple paged

Prepared (need to be prepared before use)

Prepared paged

Batch

To specify options for a single query create the query object and configure it:

For simple: Statement

For prepared: PreparedStatement

For batch: Batch

The easiest way to specify bound values in a query is using a tuple:

// Insert an int and text into the table
session
    .query_unpaged(
        "INSERT INTO ks.tab (a, b) VALUES(?, ?)",
        (2_i32, "some text")
    )
    .await?;

But the driver will accept anything implementing the trait SerializeRow.

Receiving results

The easiest way to read rows returned by a query is to cast each row to a tuple of values:

// Read rows containing an int and text
// Keep in mind that all results come in one response (no paging is done!),
// so the memory footprint and latency may be huge!
// To prevent that, use `Session::query_iter` or `Session::query_single_page`.
let query_rows = session
    .query_unpaged("SELECT a, b FROM ks.tab", &[])
    .await?
    .into_rows_result()?;

for row in query_rows.rows()? {
    // Parse row as int and text \
    let (int_val, text_val): (i32, &str) = row?;
}

See the book for more receiving methods

Modules

authentication

client

This module holds entities that represent the whole configurable driver’s client of the cluster. The following abstractions are involved:

cluster

This module holds entities that represent the cluster as a whole, nodes in the cluster (together with a pool of connections), the cluster’s state, and logic for ruling out specific nodes.

deserialize

Deserializing DB response containing CQL query results.

errors

This module contains various errors which can be returned by Session.

frame

observability

This module holds entities that allow observing and measuring driver’s and cluster’s behaviour. This includes:

policies

This module holds policies, which are entities that allow configuring the driver’s behaviour in various aspects. The common feature of all policies is that users can implement a policy on their own (because they simply need to implement a certain trait), allowing flexible customizability of the driver.

response

This module holds entities that represent responses to requests sent to the cluster by the driver. The following abstractions are involved:

routing

This module holds entities whose goal is to enable routing requests optimally, that is, choosing a target node and a shard such that it is a replica for given token.

serialize

Serializing bound values of a query to be sent to the DB.

statement

This module holds entities representing various kinds of CQL statements, together with their execution options. The following statements are supported:

value

Derive Macros

DeserializeRow

Derive macro for the DeserializeRow trait that generates an implementation which deserializes a row with a similar layout to the Rust struct.

DeserializeValue

Derive macro for the DeserializeValue trait that generates an implementation which deserializes a User Defined Type with the same layout as the Rust struct.

SerializeRow

Derive macro for the SerializeRow trait which serializes given Rust structure into bind markers for a CQL statement.

SerializeValue

Derive macro for the SerializeValue trait which serializes given Rust structure as a User Defined Type (UDT).