scylla::cluster

Module metadata

This module holds entities that represent the cluster metadata, which includes:

topology metadata:

Peer,

schema metadata:

Keyspace,

Strategy - replication strategy employed by a keyspace,

Table,

Column,

ColumnKind,

MaterializedView,

CQL types (re-exported from scylla-cql):

ColumnType,

NativeType,

UserDefinedType,

CollectionType,

Structs

Column

Describes a column of the table.

ColumnKindFromStrError

ColumnKind parse error

Keyspace

Describes a keyspace in the cluster.

MaterializedView

Describes a materialized view in the cluster.

Peer

Table

Describes a table in the cluster.

UserDefinedType

Definition of a user-defined type (UDT). UDT is composed of fields, each with a name and an optional value of its own type.

Enums

CollectionType

Collection variants of ColumnType. A collection is a composite type that has dynamic size, so it is possible to add and remove values to/from it.

ColumnKind

ColumnType

A type of:

NativeType

A ColumnType variants that are “simple” (non-recursive).

Strategy