scylla

Module errors

This module contains various errors which can be returned by Session.

Re-exports

pub use crate::client::pager::NextPageError;

pub use crate::client::pager::NextRowError;

pub use crate::response::query_result::FirstRowError;

pub use crate::response::query_result::IntoRowsResultError;

pub use crate::response::query_result::MaybeFirstRowError;

pub use crate::response::query_result::ResultNotRowsError;

pub use crate::response::query_result::RowsError;

pub use crate::response::query_result::SingleRowError;

pub use crate::authentication::AuthError;

Structs

BrokenConnectionError

An error indicating that a connection was broken. Possible error reasons:

ConnectionSetupRequestError

An error that occurred during connection setup request execution. It indicates that request needed to initiate a connection failed.

DeserializationError

An error indicating that a failure happened during deserialization.

MetadataFetchError

An error occurred during metadata fetch.

SerializationError

An error indicating that a failure happened during serialization.

TypeCheckError

An error indicating that a failure happened during type check.

Enums

BadKeyspaceName

Invalid keyspace name given to Session::use_keyspace()

BadQuery

Error caused by caller creating an invalid statement.

BrokenConnectionErrorKind

A reason why connection was broken.

ClusterStateTokenError

Error returned from ClusterState APIs.

ConnectionError

An error that appeared on a connection level. It indicated that connection can no longer be used and should be dropped.

ConnectionPoolError

An error that occurred when selecting a node connection to perform a request on.

ConnectionSetupRequestErrorKind

CqlAuthChallengeParseError

An error type returned when deserialization of AUTH_CHALLENGE response fails.

CqlAuthSuccessParseError

An error type returned when deserialization of AUTH_SUCCESS response fails.

CqlAuthenticateParseError

An error type returned when deserialization of AUTHENTICATE response fails.

CqlErrorParseError

An error type returned when deserialization of ERROR response fails.

CqlEventHandlingError

Failed to handle a CQL event received on a stream -1. Possible error kinds are:

CqlEventParseError

An error type returned when deserialization of EVENT response fails.

CqlRequestKind

Possible requests sent by the client.

CqlRequestSerializationError

An error that occurred during CQL request serialization.

CqlResponseKind

Possible CQL responses received from the server

CqlResponseParseError

An error type returned when deserialization of CQL server response fails.

CqlResultParseError

An error type returned when deserialization of RESULT response fails.

CqlSupportedParseError

An error type returned when deserialization of SUPPORTED response fails.

DbError

An error sent from the database in response to a query as described in the specification\

ExecutionError

Error that occurred during request execution

FrameBodyExtensionsParseError

An error returned by parse_response_body_extensions.

FrameHeaderParseError

An error that occurred during frame header deserialization.

KeyspaceStrategyError

An error that occurred during specific keyspace’s metadata fetch.

KeyspacesMetadataError

An error that occurred during keyspaces metadata fetch.

MetadataError

An error that occurred during metadata fetch and verification.

MetadataFetchErrorKind

Specific reason why metadata fetch failed.

NewSessionError

Error that occurred during session creation

OperationType

Type of the operation rejected by rate limiting

PagerExecutionError

An error that occurred during construction of QueryPager.

PeersMetadataError

An error that occurred during peers metadata fetch.

PrepareError

An error returned by Session::prepare().

RequestAttemptError

An error that occurred during a single attempt of:

RequestError

An error that occurred during execution of

SchemaAgreementError

An error that occurred when awating schema agreement.

TablesMetadataError

An error that occurred during tables metadata fetch.

TlsError

A wrapper around a TLS error.

TracingError

An error that occurred during tracing info fetch.

TranslationError

Error caused by failed address translation done before establishing connection

UdtMetadataError

An error that occurred during UDTs metadata fetch.

UseKeyspaceError

An error that occurred during USE KEYSPACE <> request.

WriteType

Type of write operation requested