scylla::frame
Module value

Structs

Counter

Represents an counter value

CqlDate

Native CQL date representation that allows for a bigger range of dates (-262145-1-1 to 262143-12-31).

CqlDecimal

Native CQL decimal representation.

CqlDuration

Represents a CQL Duration value

CqlTime

Native CQL time representation.

CqlTimestamp

Native CQL timestamp representation that allows full supported timestamp range.

CqlTimeuuid

Represents timeuuid (uuid V1) value

CqlVarint

Native CQL varint representation.

LegacyBatchValuesFirstSerializedDeprecated

Allows reusing already-serialized first value

LegacyBatchValuesFromIterDeprecated

Implements BatchValues from an Iterator over references to things that implement ValueList

LegacyBatchValuesIteratorFromIteratorDeprecated

Implements BatchValuesIterator from an Iterator over references to things that implement ValueList

LegacySerializedValuesDeprecated

Keeps a buffer with serialized Values Allows adding new Values and iterating over serialized ones

LegacySerializedValuesIteratorDeprecated

TupleValuesIterDeprecated

Unset

Represents an unset value

ValueOverflow

ValueTooBigDeprecated

Enums

MaybeUnset

Enum providing a way to represent a value that might be unset

SerializeValuesErrorDeprecated

Traits

LegacyBatchValuesDeprecated

Represents List of ValueList for Batch statement

LegacyBatchValuesIteratorDeprecated

An iterator-like for ValueList

ValueDeprecated

Every value being sent in a query must implement this trait serialize() should write the Value as bytes to the provided buffer

ValueListDeprecated

Represents list of values to be sent in a query gets serialized and but into request

Type Aliases

SerializedResultDeprecated