scylla
Module macros

Macros

impl_from_cql_value_from_methodDeprecated

This macro implements FromCqlVal given a type and method of CqlValue that returns this type.

impl_serialize_row_via_value_listDeprecated

Implements the SerializeRow trait for a type, provided that the type already implements the legacy ValueList trait.

impl_serialize_value_via_valueDeprecated

Implements the SerializeValue trait for a type, provided that the type already implements the legacy Value trait.

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

BytesMut

A unique reference to a contiguous slice of memory.

Traits

BufMut

A trait for values that provide sequential write access to bytes.

Derive Macros

DeserializeRow

Derive macro for the DeserializeRow trait that generates an implementation which deserializes a row with a similar layout to the Rust struct.

DeserializeValue

Derive macro for the DeserializeValue trait that generates an implementation which deserializes a User Defined Type with the same layout as the Rust struct.

FromRow

Derive macro for the FromRow trait which deserializes a row to given Rust structure.

FromUserType

#[derive(FromUserType)] allows to parse struct as a User Defined Type

IntoUserType

#[derive(IntoUserType)] allows to pass struct a User Defined Type Value in queries

SerializeRow

Derive macro for the SerializeRow trait which serializes given Rust structure into bind markers for a CQL statement.

SerializeValue

Derive macro for the SerializeValue trait which serializes given Rust structure as a User Defined Type (UDT).

ValueList

#[derive(ValueList)] allows to pass struct as a list of values for a query