Trait scylla::serialize::batch::BatchValuesIterator

pub trait BatchValuesIterator<'bv> {
    // Required methods
    fn serialize_next(
        &mut self,
        ctx: &RowSerializationContext<'_>,
        writer: &mut RowWriter<'_>,
    ) -> Option<Result<(), SerializationError>>;
    fn is_empty_next(&mut self) -> Option<bool>;
    fn skip_next(&mut self) -> Option<()>;

    // Provided method
    fn count(self) -> usize
       where Self: Sized { ... }
}

An Iterator-like object over the values from the parent BatchValues object.

It’s not a true Iterator because it does not provide direct access to the items being iterated over, instead it allows calling methods of the underlying SerializeRow trait while advancing the iterator.

Required Methods

fn serialize_next( &mut self, ctx: &RowSerializationContext<'_>, writer: &mut RowWriter<'_>, ) -> Option<Result<(), SerializationError>>

Serializes the next set of values in the sequence and advances the iterator.

fn is_empty_next(&mut self) -> Option<bool>

Returns whether the next set of values is empty or not and advances the iterator.

fn skip_next(&mut self) -> Option<()>

Skips the next set of values.

Provided Methods

fn count(self) -> usize
where Self: Sized,

Return the number of sets of values, consuming the iterator in the process.