scylla
Module transport

Re-exports

pub use execution_profile::ExecutionProfile;

Modules

downgrading_consistency_retry_policy

errors

This module contains various errors which can be returned by Session.

execution_profile

ExecutionProfile is a grouping of configurable options regarding query execution.

host_filter

Host filters.

iterator

Iterators over rows returned by paged queries

legacy_query_result

load_balancing

Load balancing configurations
Session can use any load balancing policy which implements the LoadBalancingPolicy trait
See the book for more information

locator

partitioner

query_result

retry_policy

Query retries configurations
To decide when to retry a query the Session can use any object which implements the RetryPolicy trait

session

Session is the main object used in the driver.
It manages all connections to the cluster and allows to perform queries.

session_builder

SessionBuilder provides an easy way to create new Sessions

speculative_execution

topology

Structs

ClusterData

Node

Node represents a cluster node along with it’s data and connections

PagingState

The state of a paged query, i.e. where to resume fetching result rows upon next request.

SelfIdentity

Driver and application self-identifying information, to be sent in STARTUP message.

Enums

Authenticator

Compression

The wire protocol compression algorithm.

KnownNode

Describes a database server known on Session startup.

NodeAddr

This enum is introduced to support address translation only upon opening a connection, as well as to cope with a bug present in older Cassandra and Scylla releases. The bug involves misconfiguration of rpc_address and/or broadcast_rpc_address in system.local to 0.0.0.0. Mitigation involves replacing the faulty address with connection’s address, but then that address must not be subject to AddressTranslator, so we carry that information using this enum. Address translation is never performed on Untranslatable variant.

PagingStateResponse

Type Aliases

NodeRef

A way that Nodes are often passed and accessed in the driver’s code.