Module scylla::transport::load_balancing

Load balancing configurations
Session can use any load balancing policy which implements the LoadBalancingPolicy trait
See the book for more information

Structs

DefaultPolicy

The default load balancing policy.

DefaultPolicyBuilder

The intended way to instantiate the DefaultPolicy.

LatencyAwarenessBuilder

The builder of LatencyAwareness module of DefaultPolicy.

Plan

The list of nodes constituting the query plan.

RoutingInfo

Represents info about statement that can be used by load balancing policies.

Traits

LoadBalancingPolicy

Policy that decides which nodes to contact for each query.

Type Aliases

FallbackPlan

The fallback list of nodes in the query plan.