scylla::transport
Module retry_policy

Query retries configurations
To decide when to retry a query the Session can use any object which implements the RetryPolicy trait

Structs

DefaultRetryPolicy

Default retry policy - retries when there is a high chance that a retry might help.
Behaviour based on DataStax Java Driver

DefaultRetrySession

FallthroughRetryPolicy

Forwards all errors directly to the user, never retries

FallthroughRetrySession

QueryInfo

Information about a failed query

Enums

RetryDecision

Traits

RetryPolicy

Specifies a policy used to decide when to retry a query

RetrySession

Used throughout a single query to decide when to retry it After this query is finished it is destroyed or reset