scylla

Module value

Structs

Counter

Represents an counter value

CqlDate

Native CQL date representation that allows for a bigger range of dates (-262145-1-1 to 262143-12-31).

CqlDecimal

Native CQL decimal representation.

CqlDecimalBorrowed

Borrowed version of native CQL decimal representation.

CqlDuration

Represents a CQL Duration value

CqlTime

Native CQL time representation.

CqlTimestamp

Native CQL timestamp representation that allows full supported timestamp range.

CqlTimeuuid

Represents timeuuid (uuid V1) value

CqlVarint

Native CQL varint representation.

CqlVarintBorrowed

A borrowed version of native CQL varint representation.

Row

A row in a CQL result set, containing a vector of columns. Each column can be either a CqlValue or None if the column is null.

Unset

Represents an unset value

ValueOverflow

Error type indicating that the value is too large to fit in the destination type.

Enums

CqlValue

Represents all possible CQL values that can be returned by the database.

MaybeUnset

Enum providing a way to represent a value that might be unset